SWMS_project/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.9.0
sqlparse==0.5.3
asgiref==3.8.1
tzdata==2025.2
pillow==11.2.1
numpy==1.26.4
opencv-python==4.9.0.80
twilio==9.6.3
